main.js

+++ Crud_Marçal/main.js
@@ -7,13 +7,15 @@ const closeModal = () => {
     clearFields()
     document.getElementById('modal').classList.remove('active')
 }
+
+
 const getLocalStorage = () => JSON.parse(localStorage.getItem('db_Remedio')) ?? [ ]
-const setLocalStorage = (dbRemedio) => localStorage.setItem('db_Remedio', JSON.stringify(dbRemedio))
+const setLocalStorage = (dbRemedio) => localStorage.setItem("db_Remedio", JSON.stringify(dbRemedio))
 
 //Crud aqui - Create, Read, Update, Delete
 const deleteRemedio = (index) =>{
     const dbRemedio = readRemedio()
-    dbRemedio.splice(index,1)
+    dbRemedio.splice(index, 1)
     setLocalStorage(dbRemedio)
 }
 
@@ -36,12 +38,16 @@ const isValidFields = () => {
 }
 
 //Interação com o usuário
-const clearFields =() => {
+
+const clearFields = () => {
     const fields = document.querySelectorAll('.modal-field')
-    fields.forEach(field => field.value = " ")
+    fields.forEach(field => field.value = "")
+    document.getElementById('nomedoMedicamento').dataset.index = 'new'
+    document.querySelector(".modal-header>h2").textContent  = 'Novo Remedio'
 }
+
 const saveRemedio = () => {
-    if(isValidFields()) {
+    if (isValidFields()) {
         const Remedio = {
             nomeDoMedicamento: document.getElementById('nomeDoMedicamento').value,
             SacheOuCapsula: document.getElementById ('SacheOuCapsula').value,
@@ -59,6 +65,7 @@ const saveRemedio = () => {
             updateTable()
             closeModal()
         }
+    }
 }
 
 const createRow = (Remedio, index) => {
@@ -76,7 +83,7 @@ const createRow = (Remedio, index) => {
     `
     document.querySelector('#tableRemedio>tbody').appendChild(newRow)
 }
-const clearTable = () =>{
+const clearTable = () => {
     const rows = document.querySelectorAll('#tableRemedio>tbody tr')
     rows.forEach(row => row.parentNode.removeChild(row))
 }
@@ -91,29 +98,36 @@ const fillFields = (Remedio) => {
     document.getElementById('Gramas').value = Remedio.Gramas
     document.getElementById('GramasTotal').value = Remedio.GramasTotal
     document.getElementById('CRF').value = Remedio.CRF
+    document.getElementById('nomeDoMedicamento').dataset.index = Remedio.index
 }
 
 const editRemedio = (index) => {
     const Remedio = readRemedio()[index]
     Remedio.index = index
     fillFields(Remedio)
+    document.querySelector(".modal-header>h2").textContent  = `Editando ${Remedio.nomeDoMedicamento}`
     openModal()
 }
 
 const editDelete = (event) => {
-    if (event.target.type == 'button'){
+    if (event.target.type == 'button') {
 
         const [action, index] = event.target.id.split('-')
 
         if (action == 'edit'){
             editRemedio(index)
         }else{
-            deleteRemedio(index)
-            updateTable()
-        }    
+            const Remedio = readRemedio()[index]
+            const response = confirm(`Deseja realmente excluir? ${Remedio.nomeDoMedicamento}`)
+            if (response) {
+                deleteRemedio(index)
+                updateTable()
+            }  
+        }
+    }
 }
+
 updateTable()
-}
 
  //Eventos a partir daqui
 document.getElementById('cadastrarRemedio')
@@ -127,3 +141,6 @@ document.getElementById('salvar')
 
 document.querySelector('#tableRemedio>tbody')
     .addEventListener('click', editDelete)
+
+document.getElementById('cancelar')
+    .addEventListener('click', closeModal)
